Extract PPDBForm and PPDBInfo components from PPDBPage (Component Extraction)
- Created PPDBForm component (242 lines) with encapsulated form logic and validation
- Created PPDBInfo component (62 lines) for schedule, requirements, and contact sections
- Refactored PPDBPage from 308 to 42 lines (86% reduction)
- Improved Separation of Concerns: Form logic separated from page layout
- All form accessibility attributes preserved (aria-required, aria-invalid, aria-describedby)
- Grid layout for responsive design (2-column for date/place fields)
- TypeScript compilation passed (0 errors)
- ESLint passed (0 errors)
- All 1393 tests passing (2 skipped, 154 todo)
- Zero breaking changes to existing functionality
